backend/requirements.txt
agate==1.12.0
agate-dbf==0.2.3
agate-excel==0.4.1
agate-sql==0.7.2
asgiref==3.8.1
atomicwrites==1.4.1
attrs==24.1.0
babel==2.16.0
Brotli==1.1.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==2.0.12
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.0
cssselect2==0.7.0
csvkit==2.0.1
dbfread==2.0.7
defusedxml==0.8.0rc2
Django==3.2
django-colorfield==0.11.0
django-colorful==1.3
django-extensions==2.2.9
django-filter==23.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-extra-fields==3.7.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
filetype==1.2.0
flake8==7.1.1
fonttools==4.53.1
greenlet==3.1.0
gunicorn==20.0.4
html5lib==1.1
idna==3.7
importlib_metadata==8.5.0
inflection==0.5.1
iniconfig==2.0.0
isodate==0.6.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.4
leather==0.4.0
MarkupSafe==2.1.5
mccabe==0.7.0
oauthlib==3.2.2
olefile==0.47
openpyxl==3.1.5
packaging==24.1
parsedatetime==2.6
Pillow==9.1.0
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.11.0
pycodestyle==2.12.1
pycparser==2.22
pydyf==0.11.0
pyflakes==3.2.0
PyJWT==2.9.0
pyphen==0.16.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.20.0
python-slugify==8.0.4
python3-openid==3.2.0
pytimeparse==1.1.8
pytz==2024.1
PyYAML==6.0.2
requests==2.26.0
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.5.4
SQLAlchemy==2.0.34
sqlparse==0.5.1
text-unidecode==1.3
tinycss2==1.3.0
toml==0.10.2
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.19
weasyprint==54.0
webencodings==0.5.1
xlrd==2.0.1
zipp==3.20.1
zopfli==0.2.3
